requirements.txt
openai>=1.0.0
python-telegram-bot>=20.0
gspread>=5.0.0
google-auth>=2.0.0
APScheduler>=3.10.0
feedparser>=6.0.0
requests>=2.28.0
pytrends>=4.9.0
python-dotenv>=1.0.0
